Move ImageUploadField component into a separate file
